This method will use a M1-to-M2-to-1 matching.

In this method, M2 corresponds to the number of times a treat1 unit can be matched with a treat2 unit. The M1 parameter corresponds to the number of times a treat1 unit can be used in total.

Usage

OneToN(tmatch, M1 = 2, M2 = 1, ...)

Arguments

tmatch

initial results from [trimatch()] that contains all possible matches within the specified caliper.

M1

a scaler indicating the number of unique subjects in group one to retain. This applies only to the first group in the matching order.

M2

a scaler indicating the number of unique matches to retain. This applies to the first two groups in the matching order.

...

currently unused.