Returns a ggplot2 box plot of the differences.

A boxplot of differences between each pair of treatments.

Usage

boxdiff.plot(tmatch, out, plot.mean = TRUE, ordering = attr(tmatch,
  "match.order"), ci.width = 0.5)

Arguments

tmatch

the results from trimatch.

out

a vector of the outcome measure of interest.

plot.mean

logical indicating whether the means should be plotted.

ordering

specify the order for doing the paired analysis, that is analysis will be conducted as: ordering[1] - ordering[2], ordering[1] - ordering[3], and ordering[2] - ordering[3].

ci.width

the width for the confidence intervals.

Value

a ggplot2 boxplot of the differences.