Triangle plot.

Triangle plot showing the fitted values (propensity scores) for three different models.

Usage

# S3 method for class 'triangle.psa'
plot(x, point.alpha = 0.3, point.size = 1.5,
  legend.title = "Treatment", text.size = 4, draw.edges = FALSE,
  draw.segments = TRUE, edge.alpha = 0.2, edge.color = "grey",
  edge.labels = c("Model 1", "Model 2", "Model 3"), sample = c(1), ...)

Arguments

x

the results from trips.

point.alpha

alpha level for points.

point.size

point size.

legend.title

title for the legend.

text.size

text size.

draw.edges

draw edges of the triangle.

draw.segments

draw segments connecting points across two models.

edge.alpha

alpha level for edges if drawn.

edge.color

the color for edges if drawn.

edge.labels

the labels to use for each edge of the triangle.

sample

a vector of length 1 or 3 representing the sample of points to plot. The position of each element corresponds to the groups as returned by attr(tpsa,'groups'). If equal to one, all points will be plotted. Values less than one will plot a percentage of points. Values greater than one exactly that number of points will be plotted.

...

currently unused.

Value

ggplot2 figure

See also