Provides a summary of the matched triplets including analysis of outcome measure if provided.

If an outcome measure is provided this function will perform a Freidman Rank Sum Test and repeated measures ANOVA. If either test has a statistically significant difference (as determined by the value of the p parameter), a Pairwise Wilcoxon Rank Sum Test will also be provided.

Usage

# S3 method for class 'triangle.matches'
summary(object, outcome, p = 0.05, ordering = attr(object, "match.order"), ...)

Arguments

object

result of [trimatch()].

outcome

vector representing the outcome measure.

p

threshold of the p value to perform a

ordering

specify the order for doing the paired analysis, that is analysis will be conducted as: ordering[1] - ordering[2], ordering[1] - ordering[3], and ordering[2] - ordering[3].

...

parameters passed to other statistical tests.

Value

a trimatch.summary object.

See also