Results from a study examining the effects of tutoring services on course grades.

treat Treatment indicator.

Course The course id the student was enrolled in.

Grade The course grade the student earned (4=A, 3=B, 2=C, 1=D, 0=F or W).

Gender Gender of the student.

Ethnicity Ethnicity of the student, either White, Black, or Other.

Military Is the student an active military student.

ESL English second language student.

EdMother Education level of the mother (1 = did not finish high school; 2 = high school grad; 3 = some college; 4 = earned associate degree; 5 = earned baccalaureate degree; 6 = Earned Master's degree; 7 = earned doctorate).

EdFather Education level of the father (levels same as EdMother).

Age Age at the start of the course.

Employment Employment level at college enrollment (1 = No; 2 = part-time; 3 = full-time).

Income Household income level at college enrollment (1 = <25K; 2 = <35K; 3 = <45K; 4 = <55K; 5 = <70K; 6 = <85K; 7 = <100K; 8 = <120K; 9 = >120K).

Transfer Number of transfer credits at the start of the course.

GPA GPA as of the start of the course.

GradeCode Letter grade.

Level Level of the course, either Lower or Upper.

ID Randomly assigned student ID.

Format

a data frame with 17 variables.